+++ Cargo.toml
@@ -21,6 +21,7 @@ codspeed-criterion-compat = { version = "2.6.0", default-features = false }
 colored = { version = "3.0.0" }
 criterion = { version = "0.6.0", default-features = false }
 crossbeam = { version = "0.8.4" }
+ctor = "0.4.2"
 dunce = { version = "1.0.5" }
 ignore = { version = "0.4.23" }
 indicatif = { version = "0.17.8" }
