+++ crates/karva_core/Cargo.toml
@@ -28,6 +28,7 @@ ignore = { workspace = true }
 [dev-dependencies]
 tempfile = { workspace = true }
 regex = { workspace = true }
+ctor = { workspace = true }
 
 [lints]
 workspace = true
