+++ crates/karva_core/src/lib.rs
@@ -8,4 +8,5 @@ pub mod runner;
 pub mod tag;
 pub mod utils;
 
+#[cfg(test)]
 pub mod testing;
